+++ tests/test_io_stac.py
@@ -4,6 +4,7 @@ import json
 import os
 from unittest.mock import patch
 
+import numpy
 import pytest
 import rasterio
 
@@ -138,11 +139,17 @@ def test_tile_valid(rio):
         img = stac.tile(71, 102, 8, assets="green")
         assert img.data.shape == (1, 256, 256)
         assert img.mask.shape == (256, 256)
-        assert img.band_names == ["green_1"]
+        assert img.band_names == ["green_b1"]
 
-        data, mask = stac.tile(71, 102, 8, assets=("green",))
-        assert data.shape == (1, 256, 256)
-        assert mask.shape == (256, 256)
+        img = stac.tile(71, 102, 8, assets=("green",))
+        assert img.data.shape == (1, 256, 256)
+        assert img.mask.shape == (256, 256)
+        assert img.band_names == ["green_b1"]
+
+        img = stac.tile(71, 102, 8, assets=("green", "red"))
+        assert img.data.shape == (2, 256, 256)
+        assert img.mask.shape == (256, 256)
+        assert img.band_names == ["green_b1", "red_b1"]
 
         img = stac.tile(71, 102, 8, expression="green/red")
         assert img.data.shape == (1, 256, 256)
@@ -170,7 +177,7 @@ def test_tile_valid(rio):
         )
         assert img.data.shape == (3, 256, 256)
         assert img.mask.shape == (256, 256)
-        assert img.band_names == ["green_1", "green_1", "red_1"]
+        assert img.band_names == ["green_b1", "green_b1", "red_b1"]
 
         # check backward compatibility for `indexes`
         img = stac.tile(
@@ -178,11 +185,11 @@ def test_tile_valid(rio):
             102,
             8,
             assets=("green", "red"),
-            indexes=1,
+            indexes=(1, 1),
         )
-        assert img.data.shape == (2, 256, 256)
+        assert img.data.shape == (4, 256, 256)
         assert img.mask.shape == (256, 256)
-        assert img.band_names == ["green_1", "red_1"]
+        assert img.band_names == ["green_b1", "green_b1", "red_b1", "red_b1"]
 
         img = stac.tile(
             71,
@@ -194,6 +201,30 @@ def test_tile_valid(rio):
         assert img.data.shape == (3, 256, 256)
         assert img.mask.shape == (256, 256)
         assert img.band_names == ["green_b1*2", "green_b1", "red_b1*2"]
+
+        # Check that we can use expression and asset_expression
+        # internally we will translate asset_expression b1*2 to b1 and b1+100 to b1
+        img = stac.tile(
+            71,
+            102,
+            8,
+            expression="green/red",
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img.data.shape == (1, 256, 256)
+        assert img.mask.shape == (256, 256)
+        assert img.band_names == ["green/red"]
+        img_check = stac.tile(
+            71,
+            102,
+            8,
+            assets=("green", "red"),
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img_check.data.shape == (2, 256, 256)
+        assert img_check.mask.shape == (256, 256)
+        assert img_check.band_names == ["green_b1*2", "red_b1+100"]
+        numpy.array_equal(img.data, img_check.data[0] / img_check.data[1])
 
 
 @patch("rio_tiler.io.cogeo.rasterio")
@@ -214,20 +245,20 @@ def test_part_valid(rio):
         img = stac.part(bbox, assets="green")
         assert img.data.shape == (1, 73, 83)
         assert img.mask.shape == (73, 83)
-        assert img.band_names == ["green_1"]
+        assert img.band_names == ["green_b1"]
 
-        data, mask = stac.part(bbox, assets=("green",))
-        assert data.shape == (1, 73, 83)
-        assert mask.shape == (73, 83)
+        img = stac.part(bbox, assets=("green",))
+        assert img.data.shape == (1, 73, 83)
+        assert img.mask.shape == (73, 83)
 
         img = stac.part(bbox, expression="green/red")
         assert img.data.shape == (1, 73, 83)
         assert img.mask.shape == (73, 83)
         assert img.band_names == ["green/red"]
 
-        data, mask = stac.part(bbox, assets="green", max_size=30)
-        assert data.shape == (1, 27, 30)
-        assert mask.shape == (27, 30)
+        img = stac.part(bbox, assets="green", max_size=30)
+        assert img.data.shape == (1, 27, 30)
+        assert img.mask.shape == (27, 30)
 
         with pytest.warns(ExpressionMixingWarning):
             img = stac.part(bbox, assets=("green", "red"), expression="green/red")
@@ -247,12 +278,12 @@ def test_part_valid(rio):
         )
         assert img.data.shape == (3, 73, 83)
         assert img.mask.shape == (73, 83)
-        assert img.band_names == ["green_1", "green_1", "red_1"]
+        assert img.band_names == ["green_b1", "green_b1", "red_b1"]
 
         img = stac.part(bbox, assets=("green", "red"), indexes=1)
         assert img.data.shape == (2, 73, 83)
         assert img.mask.shape == (73, 83)
-        assert img.band_names == ["green_1", "red_1"]
+        assert img.band_names == ["green_b1", "red_b1"]
 
         img = stac.part(
             bbox,
@@ -262,6 +293,25 @@ def test_part_valid(rio):
         assert img.data.shape == (3, 73, 83)
         assert img.mask.shape == (73, 83)
         assert img.band_names == ["green_b1*2", "green_b1", "red_b1*2"]
+
+        # Check that we can use expression and asset_expression
+        img = stac.part(
+            bbox,
+            expression="green/red",
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img.data.shape == (1, 73, 83)
+        assert img.mask.shape == (73, 83)
+        assert img.band_names == ["green/red"]
+        img_check = stac.part(
+            bbox,
+            assets=("green", "red"),
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img_check.data.shape == (2, 73, 83)
+        assert img_check.mask.shape == (73, 83)
+        assert img_check.band_names == ["green_b1*2", "red_b1+100"]
+        numpy.array_equal(img.data, img_check.data[0] / img_check.data[1])
 
 
 @patch("rio_tiler.io.cogeo.rasterio")
@@ -280,11 +330,11 @@ def test_preview_valid(rio):
         img = stac.preview(assets="green")
         assert img.data.shape == (1, 259, 255)
         assert img.mask.shape == (259, 255)
-        assert img.band_names == ["green_1"]
+        assert img.band_names == ["green_b1"]
 
-        data, mask = stac.preview(assets=("green",))
-        assert data.shape == (1, 259, 255)
-        assert mask.shape == (259, 255)
+        img = stac.preview(assets=("green",))
+        assert img.data.shape == (1, 259, 255)
+        assert img.mask.shape == (259, 255)
 
         img = stac.preview(expression="green/red")
         assert img.data.shape == (1, 259, 255)
@@ -308,12 +358,12 @@ def test_preview_valid(rio):
         )
         assert img.data.shape == (3, 259, 255)
         assert img.mask.shape == (259, 255)
-        assert img.band_names == ["green_1", "green_1", "red_1"]
+        assert img.band_names == ["green_b1", "green_b1", "red_b1"]
 
         img = stac.preview(assets=("green", "red"), indexes=1)
         assert img.data.shape == (2, 259, 255)
         assert img.mask.shape == (259, 255)
-        assert img.band_names == ["green_1", "red_1"]
+        assert img.band_names == ["green_b1", "red_b1"]
 
         img = stac.preview(
             assets=("green", "red"),
@@ -322,6 +372,23 @@ def test_preview_valid(rio):
         assert img.data.shape == (3, 259, 255)
         assert img.mask.shape == (259, 255)
         assert img.band_names == ["green_b1*2", "green_b1", "red_b1*2"]
+
+        # Check that we can use expression and asset_expression
+        img = stac.preview(
+            expression="green/red",
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img.data.shape == (1, 259, 255)
+        assert img.mask.shape == (259, 255)
+        assert img.band_names == ["green/red"]
+        img_check = stac.preview(
+            assets=("green", "red"),
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img_check.data.shape == (2, 259, 255)
+        assert img_check.mask.shape == (259, 255)
+        assert img_check.band_names == ["green_b1*2", "red_b1+100"]
+        numpy.array_equal(img.data, img_check.data[0] / img_check.data[1])
 
 
 @patch("rio_tiler.io.cogeo.rasterio")
@@ -342,6 +409,9 @@ def test_point_valid(rio):
 
         data = stac.point(-80.477, 33.4453, assets=("green",))
         assert len(data) == 1
+
+        data = stac.point(-80.477, 33.4453, assets=("green", "red"))
+        assert len(data) == 2
 
         data = stac.point(-80.477, 33.4453, expression="green/red")
         assert len(data) == 1
@@ -382,6 +452,24 @@ def test_point_valid(rio):
         assert len(data[0]) == 2
         assert len(data[1]) == 1
 
+        # Check that we can use expression and asset_expression
+        data = stac.point(
+            -80.477,
+            33.4453,
+            expression="green/red",
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert len(data) == 1
+
+        data_check = stac.point(
+            -80.477,
+            33.4453,
+            assets=("green", "red"),
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert len(data_check) == 2
+        assert data[0][0] == data_check[0][0] / data_check[1][0]
+
 
 @patch("rio_tiler.io.cogeo.rasterio")
 def test_statistics_valid(rio):
@@ -400,11 +488,11 @@ def test_statistics_valid(rio):
 
         stats = stac.statistics(assets="green")
         assert stats["green"]
-        assert isinstance(stats["green"]["1"], BandStatistics)
+        assert isinstance(stats["green"]["b1"], BandStatistics)
 
         stats = stac.statistics(assets=("green", "red"), hist_options={"bins": 20})
         assert len(stats) == 2
-        assert len(stats["green"]["1"]["histogram"][0]) == 20
+        assert len(stats["green"]["b1"]["histogram"][0]) == 20
 
         # Check that asset_expression is passed
         stats = stac.statistics(
@@ -419,14 +507,14 @@ def test_statistics_valid(rio):
             assets=("green", "red"), asset_indexes={"green": 1, "red": 1}
         )
         assert stats["green"]
-        assert isinstance(stats["green"]["1"], BandStatistics)
-        assert isinstance(stats["red"]["1"], BandStatistics)
+        assert isinstance(stats["green"]["b1"], BandStatistics)
+        assert isinstance(stats["red"]["b1"], BandStatistics)
 
         # Check that asset_indexes is passed
         stats = stac.statistics(assets=("green", "red"), indexes=1)
         assert stats["green"]
-        assert isinstance(stats["green"]["1"], BandStatistics)
-        assert isinstance(stats["red"]["1"], BandStatistics)
+        assert isinstance(stats["green"]["b1"], BandStatistics)
+        assert isinstance(stats["red"]["b1"], BandStatistics)
 
 
 @patch("rio_tiler.io.cogeo.rasterio")
@@ -438,23 +526,23 @@ def test_merged_statistics_valid(rio):
         with pytest.warns(UserWarning):
             stats = stac.merged_statistics()
             assert len(stats) == 3
-            assert isinstance(stats["red_1"], BandStatistics)
-            assert stats["red_1"]
-            assert stats["green_1"]
-            assert stats["blue_1"]
+            assert isinstance(stats["red_b1"], BandStatistics)
+            assert stats["red_b1"]
+            assert stats["green_b1"]
+            assert stats["blue_b1"]
 
         with pytest.raises(InvalidAssetName):
             stac.merged_statistics(assets="vert")
 
         stats = stac.merged_statistics(assets="green")
-        assert isinstance(stats["green_1"], BandStatistics)
+        assert isinstance(stats["green_b1"], BandStatistics)
 
         stats = stac.merged_statistics(
             assets=("green", "red"), hist_options={"bins": 20}
         )
         assert len(stats) == 2
-        assert len(stats["green_1"]["histogram"][0]) == 20
-        assert len(stats["red_1"]["histogram"][0]) == 20
+        assert len(stats["green_b1"]["histogram"][0]) == 20
+        assert len(stats["red_b1"]["histogram"][0]) == 20
 
         # Check that asset_expression is passed
         stats = stac.merged_statistics(
@@ -467,8 +555,8 @@ def test_merged_statistics_valid(rio):
         stats = stac.merged_statistics(
             assets=("green", "red"), asset_indexes={"green": 1, "red": 1}
         )
-        assert isinstance(stats["green_1"], BandStatistics)
-        assert isinstance(stats["red_1"], BandStatistics)
+        assert isinstance(stats["green_b1"], BandStatistics)
+        assert isinstance(stats["red_b1"], BandStatistics)
 
         # Check Expression
         stats = stac.merged_statistics(expression="green/red")
@@ -506,7 +594,7 @@ def test_info_valid(rio):
 
 
 def test_parse_expression():
-    """."""
+    """Parse assets expressions."""
     with STACReader(STAC_PATH) as stac:
         assert sorted(stac.parse_expression("green*red+red/blue+2.0")) == [
             "blue",
@@ -553,20 +641,20 @@ def test_feature_valid(rio):
         img = stac.feature(feat, assets="green")
         assert img.data.shape == (1, 118, 96)
         assert img.mask.shape == (118, 96)
-        assert img.band_names == ["green_1"]
+        assert img.band_names == ["green_b1"]
 
-        data, mask = stac.feature(feat, assets=("green",))
-        assert data.shape == (1, 118, 96)
-        assert mask.shape == (118, 96)
+        img = stac.feature(feat, assets=("green",))
+        assert img.data.shape == (1, 118, 96)
+        assert img.mask.shape == (118, 96)
 
         img = stac.feature(feat, expression="green/red")
         assert img.data.shape == (1, 118, 96)
         assert img.mask.shape == (118, 96)
         assert img.band_names == ["green/red"]
 
-        data, mask = stac.feature(feat, assets="green", max_size=30)
-        assert data.shape == (1, 30, 25)
-        assert mask.shape == (30, 25)
+        img = stac.feature(feat, assets="green", max_size=30)
+        assert img.data.shape == (1, 30, 25)
+        assert img.mask.shape == (30, 25)
 
         with pytest.warns(ExpressionMixingWarning):
             img = stac.feature(feat, assets=("green", "red"), expression="green/red")
@@ -578,12 +666,12 @@ def test_feature_valid(rio):
         )
         assert img.data.shape == (3, 118, 96)
         assert img.mask.shape == (118, 96)
-        assert img.band_names == ["green_1", "green_1", "red_1"]
+        assert img.band_names == ["green_b1", "green_b1", "red_b1"]
 
         img = stac.feature(feat, assets=("green", "red"), indexes=1)
         assert img.data.shape == (2, 118, 96)
         assert img.mask.shape == (118, 96)
-        assert img.band_names == ["green_1", "red_1"]
+        assert img.band_names == ["green_b1", "red_b1"]
 
         img = stac.feature(
             feat,
@@ -593,6 +681,16 @@ def test_feature_valid(rio):
         assert img.data.shape == (3, 118, 96)
         assert img.mask.shape == (118, 96)
         assert img.band_names == ["green_b1*2", "green_b1", "red_b1*2"]
+
+        # Check that we can use expression and asset_expression
+        img = stac.feature(
+            feat,
+            expression="green/red",
+            asset_expression={"green": "b1*2", "red": "b1+100"},
+        )
+        assert img.data.shape == (1, 118, 96)
+        assert img.mask.shape == (118, 96)
+        assert img.band_names == ["green/red"]
 
 
 def test_relative_assets():